% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Andropogon_shape-data.r
\docType{data}
\name{Andropogon_shape}
\alias{Andropogon_shape}
\title{Number of detection of Andropogon gerardii and it's target group in several counties in the Midwest}
\format{An object of class \code{SpatialPolygonsDataFrame} with 617 rows and 3 columns.}
\usage{
Andropogon_shape
}
\description{
Number of detection of Andropogon gerardii and it's target group in several counties in the Midwest
}
\keyword{datasets}
