\name{bayesLopod-package}
\alias{bayesLopod-package}
\alias{bayesLopod}
\docType{package}
\title{
\packageTitle{bayesLopod}
}
\description{
\packageDescription{bayesLopod}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{bayesLopod}
\packageIndices{bayesLopod}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{bayesLopod}

Maintainer: \packageMaintainer{bayesLopod}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
