% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopodDens.R
\name{lopodDens}
\alias{lopodDens}
\title{Kernel density estimates of global occupancy model parameters.}
\usage{
lopodDens(LopodModel, params = NULL)
}
\arguments{
\item{LopodModel}{A LopodModel object}

\item{params}{Parameters to be plotted. Default is NULL, which plots all global parameters}
}
\value{
A ggplot object.
}
\description{
Kernel density estimates of global occupancy model parameters.
}
\examples{
lopodDens(LopodModel)
}
