% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LopodModel_class.R
\docType{class}
\name{LopodModel-class}
\alias{LopodModel-class}
\alias{LopodModel}
\title{An S4 class containing a LopodModel.}
\description{
An S4 class containing a LopodModel.
}
\section{Slots}{

\describe{
\item{\code{LopodData}}{The LopodData object used to buid the LopodModel}

\item{\code{StanFit}}{stanfit object with the run model}

\item{\code{modelInfo}}{List of settings used to run the LopodModel}
}}

