% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopodSummary.R
\name{lopodSummary}
\alias{lopodSummary}
\title{Summary statistics for a LopodModel}
\usage{
lopodSummary(LopodModel, params = NULL, probs = c(0.05, 0.5, 0.95))
}
\arguments{
\item{LopodModel}{A LopodModel object}

\item{params}{Parameters to be plotted. Default is NULL, which plots all global parameters}

\item{probs}{Quantiles to be estimated}
}
\value{
Returns a dataframe which contain summaries for for all chains merged for the Global Parameters of a LopodModel. Included in the summary are quantiles, means, standard deviations (sd), effective sample sizes (n_eff), Monte Carlo standard errors (se_mean) and Rhats.
}
\description{
Summary statistics for a LopodModel
}
\examples{
lopodSummary(LopodModel)
lopodSummary(LopodModel, probs = NULL)
}
