% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopodTrace.R
\name{lopodTrace}
\alias{lopodTrace}
\title{Trace}
\usage{
lopodTrace(LopodModel, params = "lp__", inc_warmup = FALSE)
}
\arguments{
\item{LopodModel}{A LopodModel object}

\item{params}{Parameters to be plotted. Default is NULL, which plots all global parameters}
}
\value{
A ggplot object.
}
\description{
Trace
}
\examples{
lopodDens(LopodModel)
}
