% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLopodRaster-data.R
\docType{data}
\name{mLopodRaster}
\alias{mLopodRaster}
\title{bayesLopod model for a simulated species in a geographical raster}
\format{An object of class \code{LopodModel} of length 1.}
\usage{
mLopodRaster
}
\description{
bayesLopod model for a simulated species in a geographical raster
}
\keyword{datasets}
