% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mLopodShape-data.R
\docType{data}
\name{mLopodShape}
\alias{mLopodShape}
\title{bayesLopod model for Andropogon gerardii in a Shapefile}
\format{An object of class \code{LopodModel} of length 1.}
\usage{
mLopodShape
}
\description{
bayesLopod model for Andropogon gerardii in a Shapefile
}
\keyword{datasets}
