% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLopod.R
\name{modelLopod}
\alias{modelLopod}
\title{Run a  Bayes inference of Landscape Occupancy from Presence-Only Data on Stan}
\usage{
modelLopod(LopodData, varP = F, q = NULL, pmin = 0, CAR = F,
  nChains = 4, warmup = 2000, sampling = 1000, nCores = 4)
}
\arguments{
\item{LopodData}{\linkS4class{LopodData} with the data to be used in the Model}

\item{varP}{Boolean. If TRUE, detectability will vary across cells. If FALSE a global value for detectability will be estimated.}

\item{q}{Number between 0 and 1 or NULL. Rate of false detections. If NULL the values will be estimated by the model.}

\item{pmin}{Number between 0 and 1. Minimum value for detectability in a unit in which the species occurs.}

\item{CAR}{Boolean. If TRUE, (and if a adjacency matrix is included in the LopodData object) a conditional auto-regression analysis will be performed for occupancy across units.}

\item{nChains}{Number of Markov chains used by the Stan model.}

\item{warmup}{Number of iterations for each chain to be discarded as warm-up.}

\item{sampling}{Number of iterations for each chain to be sampled (after warm-up).}

\item{nCores}{Number of cores to use when executing the chains in parallel.}
}
\value{
A \linkS4class{LopodModel} object.
}
\description{
Run a  Bayes inference of Landscape Occupancy from Presence-Only Data on Stan
}
\details{
These parameters are estimated for the entire sampling area and affect all sampling units equally. These parameters can be examined using the \code{\link{lopodDens}}, \code{\link{lopodSummary}} and \code{\link{lopodTrace}} functions in bayesLopod.
\itemize{
  \item "psi": Average occupancy across the study area.
  \item "p": (if varP = FALSE) Probability of detection for all sampling units.
  \item "pmax": (if varP = TRUE) Maximum value of p.
  \item "pmin": (if varP = TRUE) Minimum value of p.
  \item "alpha": (if CAR = TRUE) alpha parameter of the conditional autoregressive model.
  \item "tau": (if CAR = TRUE) tau parameter of the conditional autoregressive model.
  \item "chi_sq"*: (EXPERIMENTAL) chi-square statistic for model.
  \item "lLh"*: (EXPERIMENTAL) total log-likelihood for the pattern observed across all sampling units, given the parameters estimated in the model.
  \item "AIC"*: (EXPERIMENTAL) Akaike information criterion value for the model.
}

\strong{Sampled sampling-units}:
These parameters are estimated for each sampling unit that has at least one sampling event. These parameters can be retrieved to a Shapefile or Raster using the \code{\link{lopodRaster}} or \code{\link{lopodShape}}functions.
\itemize{
  \item "psi_Sampled": Probability of occurrence in each sampling unit.
  \item "pCorr": (if varP = TRUE) probability of detection in each sampling unit that is occupied. This value is corrected to include unoccupied cells, in which case p = 0.
  \item "sim_y": Simulated number of detections based on the number of sampling events and the model.
  \item "sim_true_y": Simulated number of true detections based on the number of sampling events and the model.
  \item "sim_false_y": Simulated number of false detections based on the number of sampling events and the model.
  \item "pp": Probability of presence given a pattern of sampling effort/detections and the model.
  \item "cellpres_i": Simulated presence/absence for each iteration based on model.
  \item "expRec"*: (EXPERIMENTAL) Expected number of detections.
  \item "lLh_cell"*: (EXPERIMENTAL) Likelihood of the pattern in a cell given the model.
}

\strong{All sampling-units}:
These parameters are estimated for all sampling units only if a conditional autoregressive model  is implemented (CAR = TRUE). These parameters can be retrieved to a Shapefile or Raster using the \code{\link{lopodRaster}} or \code{\link{lopodShape}} functions.
\itemize{
  \item "psi_i":  Probability of occurrence in each sampling units. This is inferred for unsampled united using the CAR model. For sampled units, "psi_i" equals "psi_Sampled"
}
*EXPERIMENTAL VALUES SHOULD BE TAKEN CAREFULLY AND NOT USED NAIVELY TO DETERMINE WHICH MODELS ARE BETTER THAN OTHERS
}
\examples{
data("simSpRecords", package = "bayesLopod")
data("simSpSamplingEffort", package = "bayesLopod")
simSpRasters = xyToRaster(xyRecords = simSpRecords,xySamplingEffort = simSpSamplingEffort,
basemap = NULL, nrows = 10, extentExpansion = 0)
ld_Raster = rasterLopodData(rasterN = simSpRasters[["samplingEffort"]],
rasterY = simSpRasters[["spDetections"]], Adjacency = FALSE )
mLopodRaster = modelLopod(LopodData = ld_Raster, varP = TRUE, q = NULL,
pmin = 0, CAR = FALSE, nChains = 1,warmup = 10,sampling = 10,nCores = 1)

data("Andropogon_shape", package = "bayesLopod")
ld_Shape = shapeLopodData(Shapefile = Andropogon_shape, fieldN = "sampEffort",
fieldY = "detections",  Adjacency = FALSE, keepFields = FALSE)
mLopodShape = modelLopod(LopodData = ld_Shape, varP = FALSE, q = NULL,
pmin = 0, CAR = FALSE, nChains = 1,warmup = 5,sampling = 5,nCores = 1)

\dontrun{
data("simSpRecords", package = "bayesLopod")
data("simSpSamplingEffort", package = "bayesLopod")
simSpRasters = xyToRaster(xyRecords = simSpRecords,xySamplingEffort = simSpSamplingEffort,
basemap = NULL, nrows = 50, extentExpansion = 0)
ld_Raster_adMatrix = rasterLopodData(rasterN = simSpRasters[["samplingEffort"]],
rasterY = simSpRasters[["spDetections"]], Adjacency = TRUE )
mLopodRaster = modelLopod(LopodData = ld_Raster_adMatrix, varP = TRUE, q = NULL,
pmin = 0.1, CAR = FALSE, nChains = 4,warmup = 500,sampling = 100,nCores = 4)

data("Andropogon_shape", package = "bayesLopod")
ld_Shape = shapeLopodData(Shapefile = Andropogon_shape, fieldN = "sampEffort",
fieldY = "detections",  Adjacency = TRUE, keepFields = FALSE)
mLopodShape = modelLopod(LopodData = ld_Shape, varP = TRUE, q = NULL,
pmin = 0, CAR = TRUE, nChains = 4,warmup = 500,sampling = 100,nCores =4)
}
}
