% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams.R
\name{modelParams}
\alias{modelParams}
\title{Internal funtion to extract parameters from a StanModel object}
\usage{
modelParams(LopodModel)
}
\description{
Internal funtion to extract parameters from a StanModel object
}
