% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterDistToRecords.R
\name{rasterDistToRecords}
\alias{rasterDistToRecords}
\title{Internal - Create a Stack with distance matrices to sampled and detected cells}
\usage{
rasterDistToRecords(rasterN, rasterY)
}
\arguments{
\item{rasterN}{Raster object with sampling effort (number of sampling events)in each cell.}

\item{rasterY}{Raster object with number of detections in each cell.}
}
\value{
A Stack object with distance to sampled cells and distance to detections.
}
\description{
Internal - Create a Stack with distance matrices to sampled and detected cells
}
