% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterStudyArea.R
\name{rasterStudyArea}
\alias{rasterStudyArea}
\title{Internal - Create a Raster object with Study Area from Raster data}
\usage{
rasterStudyArea(DistStak, maxDist, extSample = 0.025, extDetection = 0.15)
}
\arguments{
\item{DistStak}{Stack object distance to Sampled cells and detections (in that order).}

\item{maxDist}{Maximum distance in Raster}

\item{extSample}{Number between 0 and 1. Maximum distance (relative to the diagonal of the raster) from a sampled cell that should be included in the study area. If 0, there is no extrapolation to not-sampled cells.}

\item{extDetection}{Number between 0 and 1. Maximum distance (relative to the diagonal of the raster) from cell in which the species was detected that should be included in the study area. If 0, there is no extrapolation to no-sampled cells.}
}
\value{
A Raster object with Study Area.
}
\description{
Internal - Create a Raster object with Study Area from Raster data
}
