% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpRecords-data.R
\docType{data}
\name{simSpRecords}
\alias{simSpRecords}
\title{Coordinates for detections of a simulated species}
\format{An object of class \code{SpatialPoints} of length 8634.}
\usage{
simSpRecords
}
\description{
Coordinates for detections of a simulated species
}
\keyword{datasets}
