% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpSamplingEffort-data.R
\docType{data}
\name{simSpSamplingEffort}
\alias{simSpSamplingEffort}
\title{Coordinates for sampling events targeting a simulated species}
\format{An object of class \code{SpatialPoints} of length 37860.}
\usage{
simSpSamplingEffort
}
\description{
Coordinates for sampling events targeting a simulated species
}
\keyword{datasets}
