% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LopodData-class.R
\docType{class}
\name{spatialObjectsUnion-class}
\alias{spatialObjectsUnion-class}
\title{An S4 class to contain data to be input into a bayesLopod model.}
\description{
An S4 class to contain data to be input into a bayesLopod model.
}
\section{Slots}{

\describe{
\item{\code{geoDataObject}}{Spatial object supported by bayesLopod}

\item{\code{geoType}}{Type of geographical data (Only rasters supported for now)}

\item{\code{geoInfo}}{Additional spatal information to be passsed to modelLopod}
}}

