% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopodRaster.R
\name{lopodRaster}
\alias{lopodRaster}
\title{Crate raster object for a parameter estimated in a LopodModel}
\usage{
lopodRaster(LopodModel, param, extrapolate = T, metric = NULL,
  quant = 0.5)
}
\arguments{
\item{LopodModel}{A LopodModel object}

\item{param}{Unit-level model parameter to be mapped. Values "psy_Sampled" can be mapped for models without CAR analyses, "psy_i" for LopodModels with CAR analysis and "pp","cellpres_i", "pCorr","sim_y","sim_true_y","sim_false_y" for both.}

\item{extrapolate}{Boolean. If True, parameters are mapped for cells that have not been sampled, this can only be done in LopodModels with CAR analysis. Only plotted for "psy_i", "pp" and, "cellpres_i".}

\item{metric}{"mean" or "sd". Plots the mean or standard deviation of the posterior distribution. If NULL, the value in quant is used.}

\item{quant}{Returns the raster for a given quantile of the posterior distribution. Default is 0.5 (the median of the posterior distribution). Not used if metric is other than NULL}
}
\value{
A Raster object.
}
\description{
Crate raster object for a parameter estimated in a LopodModel
}
\examples{
data("mLopodRaster", package = "bayesLopod")
psyRaster = lopodRaster(mLopodRaster, param = "psy_i", extrapolate = TRUE, quant = 0.5)
ppRaster = lopodRaster(mLopodRaster, param = "pp", extrapolate = FALSE, metric = "mean")

\dontrun{
#Visualize results
sp::spplot(psyRaster)
sp::spplot(ppRaster)
}
}
