% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lopodTrace.R
\name{lopodTrace}
\alias{lopodTrace}
\title{Plots the values of model parameters for each chain across iterations}
\usage{
lopodTrace(LopodModel, params = "lp__", inc_warmup = FALSE)
}
\arguments{
\item{LopodModel}{A LopodModel object}

\item{params}{Parameters to be plotted. Default is "lp__" which plots the log posterior probability}

\item{inc_warmup}{Boolean. If true, warm-up iterations are plotted. Default is FALSE.}
}
\value{
A ggplot object.
}
\description{
Plots the values of model parameters for each chain across iterations
}
\examples{
data("mLopodRaster", package = "bayesLopod")
lopodTrace(mLopodRaster, inc_warmup = F, params = c("p","q"))

data("mLopodShape", package = "bayesLopod")
lopodTrace(mLopodShape, inc_warmup = T)
}
