% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelLopod.R
\name{modelLopod}
\alias{modelLopod}
\title{Run a  Bayes inference of Landscape Occupancy from Presence-Only Data on Stan}
\usage{
modelLopod(LopodData, varP = F, q = NULL, pmin = 0, CAR = F,
  nChains = 4, warmup = 2000, sampling = 1000, nCores = 4)
}
\arguments{
\item{LopodData}{Obejct with the data to be used in the Model}

\item{varP}{Boolean. If TRUE, detectability will vary across cells. If FALSE a global value for detectability will be estimated.}

\item{q}{Number between 0 and 1 or NULL. Rate of false detections. If NULL the values will be estimated by the model.}

\item{pmin}{Number between 0 and 1. Minimum value for detectability in a unit in which the species occurs.}

\item{CAR}{Boolean. If TRUE, (and if a adjancency matrix is included in the LopodData object) a conditional autoregression analysis will be performed for occupancy across units.}
}
\value{
A LopodModel object.
}
\description{
Run a  Bayes inference of Landscape Occupancy from Presence-Only Data on Stan
}
