% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams.R
\name{modelParams}
\alias{modelParams}
\title{Internal function to extract parameters from a StanModel object}
\usage{
modelParams(LopodModel)
}
\arguments{
\item{LopodModel}{A LopodModel object}
}
\value{
A  list of parameters.
}
\description{
Internal function to extract parameters from a StanModel object
}
